[inductor] Simplify FlyDSL FlexAttention forward kernel
